+++ src/dockerfiles/Automattic__wp-calypso-34597.Dockerfile
@@ -52,6 +52,7 @@ ENV DBUS_SESSION_BUS_ADDRESS="unix:path=/run/dbus/system_bus_socket"
 RUN dbus-daemon --system --fork
 
 ENV PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true
+ENV PUPPETEER_SKIP_DOWNLOAD=true
 ENV OPENSSL_CONF /etc/ssl
 
 RUN useradd -m chromeuser
@@ -91,43 +92,35 @@ python2 -V
 EOF_16a722167964
 
 
-RUN <<EOF_9d3989d38880
+RUN <<EOF_9b3b2b38c0ff
 #!/bin/bash
 set -euxo pipefail
 git clone -o origin https://github.com/Automattic/wp-calypso /testbed
-chmod -R 777 /testbed
 cd /testbed
 git reset --hard 6324c6d30d4ece27eeb00faa3e7cc0aa1f817f83
 git remote remove origin
-TARGET_EPOCH=$(git show -s --format=%ct 6324c6d30d4ece27eeb00faa3e7cc0aa1f817f83)
-git tag -l | while read tag; do TAG_COMMIT=$(git rev-list -n 1 "$tag"); TAG_EPOCH=$(git show -s --format=%ct "$TAG_COMMIT"); if [ "$TAG_EPOCH" -gt "$TARGET_EPOCH" ]; then git tag -d "$tag"; fi; done
-git branch -D $(git branch | grep -v "^\*") 2>/dev/null || true
+git branch | grep -v '^\*' | xargs -r git branch -D || true
+git tag -l | xargs -r git tag -d
 git reflog expire --expire=now --all
+git gc --prune=now --aggressive
+TARGET_EPOCH=$(git show -s --format=%ct 6324c6d30d4ece27eeb00faa3e7cc0aa1f817f83)
+AFTER_EPOCH=$((TARGET_EPOCH + 1))
+AFTER_TIMESTAMP=$(date -u -d "@$AFTER_EPOCH" "+%Y-%m-%d %H:%M:%S")
+COMMIT_COUNT=$(git log --oneline --all --since="$AFTER_TIMESTAMP" | wc -l)
+[ "$COMMIT_COUNT" -eq 0 ] || exit 1
 cd - || true
+chmod -R 777 /testbed
 cd /testbed
 git clean -fdxq
 source $NVM_DIR/nvm.sh
-npm install --unsafe-perm
-EOF_9d3989d38880
+sed -i 's/"color-studio": "1.0.5"/"@automattic\/color-studio": "1.0.6"/' package.json
+npm install --unsafe-perm --ignore-scripts
+npm rebuild node-sass
+ln -sf $(pwd)/node_modules/@automattic/color-studio node_modules/color-studio
+npm run build-packages
+EOF_9b3b2b38c0ff
 
 
-RUN <<EOF_959f01873f17
-#!/bin/bash
-set -euxo pipefail
-mkdir -p /swebench/image_assets
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/61068779-b8461000-a402-11e9-9a04-762503953fe8.png' 'https://user-images.githubusercontent.com/43215253/61068779-b8461000-a402-11e9-9a04-762503953fe8.png' || true
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/61068824-d6137500-a402-11e9-9eb1-4f64be224585.png' 'https://user-images.githubusercontent.com/43215253/61068824-d6137500-a402-11e9-9eb1-4f64be224585.png' || true
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/61068851-ea577200-a402-11e9-88f1-af06860dc826.png' 'https://user-images.githubusercontent.com/43215253/61068851-ea577200-a402-11e9-88f1-af06860dc826.png' || true
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/33277310-155312ba-d390-11e7-8b77-21de899376f2.png' 'https://user-images.githubusercontent.com/8658164/33277310-155312ba-d390-11e7-8b77-21de899376f2.png' || true
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/33277311-1655ab28-d390-11e7-933f-c5050e50dd96.png' 'https://user-images.githubusercontent.com/8658164/33277311-1655ab28-d390-11e7-933f-c5050e50dd96.png' || true
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/33277459-86d0d602-d390-11e7-916c-81556b74c3cf.png' 'https://user-images.githubusercontent.com/8658164/33277459-86d0d602-d390-11e7-916c-81556b74c3cf.png' || true
-EOF_959f01873f17
-
+COPY src/image_assets/Automattic__wp-calypso-34597/ /swebench/image_assets/
 
 WORKDIR /testbed
