+++ src/dockerfiles/Automattic__wp-calypso-35531.Dockerfile
@@ -52,6 +52,7 @@ ENV DBUS_SESSION_BUS_ADDRESS="unix:path=/run/dbus/system_bus_socket"
 RUN dbus-daemon --system --fork
 
 ENV PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true
+ENV PUPPETEER_SKIP_DOWNLOAD=true
 ENV OPENSSL_CONF /etc/ssl
 
 RUN useradd -m chromeuser
@@ -91,35 +92,35 @@ python2 -V
 EOF_16a722167964
 
 
-RUN <<EOF_b80f08337edc
+RUN <<EOF_b69c2f9c9599
 #!/bin/bash
 set -euxo pipefail
 git clone -o origin https://github.com/Automattic/wp-calypso /testbed
-chmod -R 777 /testbed
 cd /testbed
 git reset --hard 9e445e13dd96de61e7b7c5c98ad98e12268a4d12
 git remote remove origin
-TARGET_EPOCH=$(git show -s --format=%ct 9e445e13dd96de61e7b7c5c98ad98e12268a4d12)
-git tag -l | while read tag; do TAG_COMMIT=$(git rev-list -n 1 "$tag"); TAG_EPOCH=$(git show -s --format=%ct "$TAG_COMMIT"); if [ "$TAG_EPOCH" -gt "$TARGET_EPOCH" ]; then git tag -d "$tag"; fi; done
-git branch -D $(git branch | grep -v "^\*") 2>/dev/null || true
+git branch | grep -v '^\*' | xargs -r git branch -D || true
+git tag -l | xargs -r git tag -d
 git reflog expire --expire=now --all
+git gc --prune=now --aggressive
+TARGET_EPOCH=$(git show -s --format=%ct 9e445e13dd96de61e7b7c5c98ad98e12268a4d12)
+AFTER_EPOCH=$((TARGET_EPOCH + 1))
+AFTER_TIMESTAMP=$(date -u -d "@$AFTER_EPOCH" "+%Y-%m-%d %H:%M:%S")
+COMMIT_COUNT=$(git log --oneline --all --since="$AFTER_TIMESTAMP" | wc -l)
+[ "$COMMIT_COUNT" -eq 0 ] || exit 1
 cd - || true
+chmod -R 777 /testbed
 cd /testbed
 git clean -fdxq
 source $NVM_DIR/nvm.sh
-npm install --unsafe-perm
-EOF_b80f08337edc
+sed -i 's/"color-studio": "1.0.5"/"@automattic\/color-studio": "1.0.6"/' package.json
+npm install --unsafe-perm --ignore-scripts
+npm rebuild node-sass
+ln -sf $(pwd)/node_modules/@automattic/color-studio node_modules/color-studio
+npm run build-packages
+EOF_b69c2f9c9599
 
 
-RUN <<EOF_206d4e0ed878
-#!/bin/bash
-set -euxo pipefail
-mkdir -p /swebench/image_assets
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/63221565-9afe2180-c1b8-11e9-9a21-75af5cd75175.png' 'https://user-images.githubusercontent.com/18581859/63221565-9afe2180-c1b8-11e9-9a21-75af5cd75175.png' || true
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/63221567-9b96b800-c1b8-11e9-927f-52d47ead4789.png' 'https://user-images.githubusercontent.com/18581859/63221567-9b96b800-c1b8-11e9-927f-52d47ead4789.png' || true
-EOF_206d4e0ed878
-
+COPY src/image_assets/Automattic__wp-calypso-35531/ /swebench/image_assets/
 
 WORKDIR /testbed
